+++ wkcuber/wkcuber/upload.py
@@ -6,7 +6,7 @@ from webknossos import Dataset, webknossos_context
 from webknossos.client._defaults import DEFAULT_WEBKNOSSOS_URL
 from webknossos.client._upload_dataset import DEFAULT_SIMULTANEOUS_UPLOADS
 
-from .utils import add_verbose_flag, setup_logging
+from ._internal.utils import add_verbose_flag, setup_logging
 
 
 def create_parser() -> ArgumentParser:
